refactor: reorganize theme structure and fix final colour references
