+++ src/config/shared.ts
