+++ src/main.ts
@@ -1,5 +1,5 @@
 import "./types/theme"
 
-export { default } from "./themeLight"
-export { default as themeDark } from "./themeDark"
-export { default as themeCombined } from "./themeCombined"
+export { default } from "./themes/themeLight"
+export { default as themeDark } from "./themes/themeDark"
+export { default as themeCombined } from "./themes/themeCombined"
